// @cloud-materials/common/components/CForm/interface.ts
export type CFormStep = Omit<IFormStep, 'form'> & { form: CForm };
// ...
export type CFormFooter<T extends ObjectType = any, D extends ObjectType = any> =
  | ((form: CForm<T, D>, onSubmit: (values: T) => void) => CFormFooterConfig<T, D>)
  | CFormFooterConfig<T, D>;
export type CFormFooterConfig<T extends ObjectType = any, D extends ObjectType = any> = {
  left?: CFormFooterItem<T, D>[];
  right?: CFormFooterItem<T, D>[];
};
export type CFormFooterCustomRender = () => ReactNode;
export type CFormFooterItem<T extends ObjectType = any, D extends ObjectType = any> =
  | { field: CFieldConfig<T, D> }
  | CFormFooterCustomRender
  | CFormFooterBuiltInButton;

export type CFormStepFooter<T extends ObjectType = any, D extends ObjectType = any> =
  | ((formStep: CFormStep, onSubmit: (values: T) => void) => CFormStepFooterConfig<T, D>)
  | CFormStepFooterConfig<T, D>;
export type CFormStepFooterConfig<T extends ObjectType = any, D extends ObjectType = any> = {
  /**
   * 是否开启自动隐藏上一步/下一步/提交按钮逻辑
   * 第一步默认隐藏上一步和提交按钮
   * 最后一步默认隐藏下一步按钮
   */
  autoHiddenStepButtons?: boolean;
  left?: CFormStepFooterItem<T, D>[];
  right?: CFormStepFooterItem<T, D>[];
};
export type CFormStepFooterCustomRender = (formStep: CFormStep) => ReactNode;
export type CFormFooterBuiltInButton = {
  buttonType: 'confirm' | 'submit' | 'next' | 'back' | 'cancel';
  props?: ButtonProps;
  wrapper?: (buttonNode: React.ReactNode) => React.ReactNode;
  // 针对按钮关闭 autoHiddenStepButtons
  disableAutoHiddenStepButtons?: boolean;
};
export type CFormStepFooterItem<T extends ObjectType = any, D extends ObjectType = any> =
  | { field: CFieldConfig<T, D> }
  | CFormStepFooterCustomRender
  | CFormFooterBuiltInButton;
// ...
export interface CForm<T extends ObjectType = any, D extends ObjectType = any> extends Form<T> {
  /**
   * @zh 全局数据
   */
  data: D;
  /**
   * @zh 设置全局数据
   */
  setData: (data: Partial<D>) => void;
  /**
   * @zh 标识form名称，对应form标签的name属性
   * @default 随机的唯一uid
   * 默认取formily的form实例id属性
   */
  name: string;
}
// ...
export type ObjectType = Record<string, any>;

export type JSXComponent = keyof JSX.IntrinsicElements | React.JSXElementConstructor<any>;

export type ComponentsMap = Record<string, JSXComponent>;

export type PureComponentsMap = Record<string, React.JSXElementConstructor<any>>;